backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
python-jose
pydantic
supabase
fpdf
google-generativeai